+++ phial/bot.py
@@ -5,8 +5,8 @@ import logging
 import json
 from phial.commands import help_command
 from phial.globals import _command_ctx_stack, command, _global_ctx_stack
-from phial.wrappers import Command, Response, Message, Attachment
 from phial.scheduler import Scheduler, Schedule, ScheduledJob
+from phial.wrappers import Command, Response, Message, Attachment
 
 
 class Phial():
@@ -34,6 +34,7 @@ class Phial():
         self.config = config
         self.running = False
         self.scheduler = Scheduler()
+        self.fallback_command_func = None  # type: Optional[Callable]
         if logger is None:
             logger = logging.getLogger(__name__)
             if not logger.hasHandlers():
@@ -150,6 +151,48 @@ class Phial():
                 return m.groupdict(), command_pattern
         raise ValueError('Command "{}" has not been registered'
                          .format(text))
+
+    def fallback_command(self) -> Callable:
+        '''
+        A decorator that is used to register a command function for use
+        when a user tries to execute a command that doesn't exist
+
+        Example:
+            ::
+
+                @bot.fallback_command()
+                def error_handler(attempted_command: Command):
+                    return "Oops that command doesn't seem to exist"
+
+        '''
+        def decorator(f: Callable) -> Callable:
+            self.add_fallback_command(f)
+            return f
+        return decorator
+
+    def add_fallback_command(self, command_func: Callable) -> None:
+        '''
+        Registers a fallback function to run when a user tries to execute a
+        command that doesn't exist. This is the same as
+        :meth:`fallback_command`.
+
+        ::
+
+            @bot.fallback_command()
+            def error_handler(attempted_command: Command):
+                return "Oops that command doesn't seem to exist"
+
+        Is the same as ::
+
+            def error_handler(attempted_command: Command):
+                return "Oops that command doesn't seem to exist"
+
+            bot.add_fallback_command(error_handler)
+
+        Args:
+            command_func(func): The function to be set as the fallback
+        '''
+        self.fallback_command_func = command_func
 
     def command(self,
                 command_pattern_template: str,
@@ -340,7 +383,15 @@ class Phial():
     def _create_command(self,
                         command_message: Message) -> Optional[Command]:
         '''Creates an instance of a command'''
-        command_match = self.get_command_match(command_message.text)
+        try:
+            command_match = self.get_command_match(command_message.text)
+        except ValueError:
+            # Command not registered.
+            return Command(None,
+                           command_message.channel,
+                           None,
+                           command_message.user,
+                           command_message)
         if command_match:
             kwargs, command_pattern = command_match
             return Command(command_pattern,
@@ -354,8 +405,25 @@ class Phial():
         '''Executes a given command'''
         if command is None:
             return  # Do nothing if no command
+
+        if command.command_pattern is None:
+            command_func = None
+        else:
+            command_func = self.commands.get(command.command_pattern, None)
+
+        if command_func is None:
+            # If no command found warn and then return early
+            self.logger.warn("Command {0} not found"
+                             .format(command.message.text))
+            if self.fallback_command_func is None:
+                return
+            _command_ctx_stack.push(command)
+            return self.fallback_command_func(command)
+        if command.args is None:
+            self.logger.exception("Command has no args")
+            raise ValueError("Command has no args")
         _command_ctx_stack.push(command)
-        return self.commands[command.command_pattern](**command.args)
+        return command_func(**command.args)
 
     def _parse_slack_output(self,
                             slack_rtm_output: List[Dict])-> Optional[Message]: