+++ tests/test_bot.py
@@ -59,6 +59,26 @@ class TestCommandDecarator(TestPhialBot):
                                                 None)
 
 
+class TestFallbackCommandDecarator(TestPhialBot):
+    '''Tests for phial's fallback command decorator'''
+
+    def test_fallback_command_decorator_functionality(self):
+        @self.bot.fallback_command()
+        def command_function(command):
+            return 'oops'
+
+        self.assertEqual(self.bot.fallback_command_func,
+                         command_function)
+
+    def test_fallback_command_decorator_calls_add_fallback_command(self):
+        self.bot.add_fallback_command = MagicMock()
+
+        @self.bot.fallback_command()
+        def command_function(command):
+            return 'oops'
+        self.bot.add_fallback_command.assert_called_with(command_function)
+
+
 class TestAliasDecarator(TestPhialBot):
     '''Tests for phial's alias decorator'''
 
@@ -131,6 +151,51 @@ class TestAddCommand(TestPhialBot):
 
         self.assertEqual(self.bot.commands[expected_command_pattern]._help,
                          help_text)
+
+
+class TestFallbackCommand(TestPhialBot):
+    '''Tests for phial's add_fallback_command function'''
+
+    def test_add_fallback_command_functionality(self):
+        def command_function():
+            return 'oops'
+        self.bot.add_fallback_command(command_function)
+
+        self.assertEqual(self.bot.fallback_command_func,
+                         command_function)
+
+    def test_fallback_command_works_correctly(self):
+        def command_function(command):
+            return "oops"
+        self.bot.add_fallback_command(command_function)
+        message = phial.wrappers.Message(text="!test",
+                                         channel="channel_id",
+                                         user="user",
+                                         timestamp="timestamp")
+        command_patern = re.compile('^test$')
+        command = phial.wrappers.Command(command_patern,
+                                         'channel_id',
+                                         {},
+                                         'user',
+                                         message)
+
+        result = self.bot._handle_command(command)
+        self.assertEqual(result, "oops")
+
+    def test_no_fallback_command_does_nothing(self):
+        message = phial.wrappers.Message(text="!test",
+                                         channel="channel_id",
+                                         user="user",
+                                         timestamp="timestamp")
+        command_patern = re.compile('^test$')
+        command = phial.wrappers.Command(command_patern,
+                                         'channel_id',
+                                         {},
+                                         'user',
+                                         message)
+
+        result = self.bot._handle_command(command)
+        self.assertEqual(result, None)
 
 
 class TestBuildCommandPattern(TestPhialBot):
@@ -241,15 +306,18 @@ class TestCreateCommand(TestPhialBot):
                                                   command_message)
         self.assertEqual(command, expected_command)
 
-    def test_errors_when_no_command_match(self):
-        with self.assertRaises(ValueError) as context:
-            command_message = phial.wrappers.Message('!test',
-                                                     'channel_id',
-                                                     'user',
-                                                     'timestamp')
-            self.bot._create_command(command_message)
-        self.assertTrue('Command "test" has not been registered'
-                        in str(context.exception))
+    def test_returns_partialcommand_when_no__command_match(self):
+        command_message = phial.wrappers.Message('!test',
+                                                 'channel_id',
+                                                 'user',
+                                                 'timestamp')
+        expected_result = phial.wrappers.Command(None,
+                                                 'channel_id',
+                                                 None,
+                                                 'user',
+                                                 command_message)
+        result = self.bot._create_command(command_message)
+        self.assertEqual(result, expected_result)
 
 
 class TestHandleCommand(TestPhialBot):
@@ -272,6 +340,11 @@ class TestHandleCommand(TestPhialBot):
         self.bot._handle_command(command_instance)
 
         self.assertTrue(test_func.called)
+
+    def test_handle_command_does_nothing_when_called_with_none(self):
+        result = self.bot._handle_command(None)
+
+        self.assertEqual(result, None)
 
 
 class TestCommandContextWorksCorrectly(TestPhialBot):
@@ -845,8 +918,8 @@ class TestRun(TestPhialBot):
                                               'timestamp')
         self.bot._parse_slack_output = MagicMock(return_value=test_command)
 
-        expected_msg = 'ValueError: Command "test" has not been registered'
-        with self.assertLogs(logger='phial.bot', level='ERROR') as cm:
+        expected_msg = 'Command !test not found'
+        with self.assertLogs(logger='phial.bot', level='WARN') as cm:
             self.bot.run()
 
             error = cm.output[0]